[Feat] 아이템 사용시 수분, 포만감 회복 구현

+++ Assets/01 Scripts/Stat/Player/TimeDecreasingStat/TimeDecreasingStat.cs
@@ -31,7 +31,7 @@ public class TimeDecreasingStat : MonoBehaviour
 
     protected virtual void OnExitZeroStat() { }
 
-    public void Heal(int amount)
+    public void Heal(float amount)
     {
         _current += amount;
         OnExitZeroStat();
